+++ migration/index.ts
@@ -6,24 +6,27 @@ import {
 } from "@typescript-eslint/utils"; // eslint-disable-line allowed/dependencies -- assumed transitive dependency
 
 type NamedProp = TSESTree.PropertyNonComputedName & {
-  key: TSESTree.Identifier;
+  key: TSESTree.Identifier | TSESTree.StringLiteral;
 };
 
 interface Queries {
-  zod: TSESTree.ImportDeclaration;
-  dateInOutExample: NamedProp;
-  getExamples: TSESTree.CallExpression;
+  dependsOnMethod: TSESTree.NewExpression;
 }
 
 type Listener = keyof Queries;
 
 const queries: Record<Listener, string> = {
-  zod: `${NT.ImportDeclaration}[source.value='zod/v4']`,
-  dateInOutExample:
-    `${NT.CallExpression}[callee.object.name='ez'][callee.property.name=/date(In|Out)/] >` +
-    `${NT.ObjectExpression} > ${NT.Property}[key.name='example']`,
-  getExamples: `${NT.CallExpression}[callee.name='getExamples']`,
+  dependsOnMethod: `${NT.NewExpression}[callee.name='DependsOnMethod']`,
 };
+
+const isNamedProp = (prop: TSESTree.ObjectLiteralElement): prop is NamedProp =>
+  prop.type === NT.Property &&
+  !prop.computed &&
+  (prop.key.type === NT.Identifier ||
+    (prop.key.type === NT.Literal && typeof prop.key.value === "string"));
+
+const getPropName = (prop: NamedProp): string =>
+  prop.key.type === NT.Identifier ? prop.key.name : prop.key.value;
 
 const listen = <
   S extends { [K in Listener]: TSESLint.RuleFunction<Queries[K]> },
@@ -38,7 +41,7 @@ const listen = <
     {},
   );
 
-const v25 = ESLintUtils.RuleCreator.withoutDocs({
+const v26 = ESLintUtils.RuleCreator.withoutDocs({
   meta: {
     type: "problem",
     fixable: "code",
@@ -53,42 +56,58 @@ const v25 = ESLintUtils.RuleCreator.withoutDocs({
   defaultOptions: [],
   create: (ctx) =>
     listen({
-      zod: (node) =>
+      dependsOnMethod: (node) => {
+        if (node.arguments.length !== 1) return;
+        const argument = node.arguments[0];
+        if (argument.type !== NT.ObjectExpression) return;
+        let isDeprecated = false;
+        let nested: TSESTree.ObjectExpression | undefined = undefined;
+        let cursor: TSESTree.Node = node;
+        while (
+          cursor &&
+          cursor.parent &&
+          cursor.parent.type === NT.MemberExpression &&
+          cursor.parent.property.type === NT.Identifier &&
+          cursor.parent.parent &&
+          cursor.parent.parent.type === NT.CallExpression
+        ) {
+          const name = cursor.parent.property.name;
+          const call = cursor.parent.parent as TSESTree.CallExpression;
+          if (name === "deprecated") isDeprecated = true;
+          if (
+            name === "nest" &&
+            call.arguments[0] &&
+            call.arguments[0].type === NT.ObjectExpression
+          )
+            nested = call.arguments[0];
+          cursor = call;
+        }
         ctx.report({
-          node: node.source,
-          messageId: "change",
-          data: { subject: "import", from: "zod/v4", to: "zod" },
-          fix: (fixer) => fixer.replaceText(node.source, `"zod"`),
-        }),
-      dateInOutExample: (node) =>
-        ctx.report({
-          node,
-          messageId: "change",
-          data: { subject: "property", from: "example", to: "examples" },
-          fix: (fixer) =>
-            fixer.replaceText(
-              node,
-              `examples: [${ctx.sourceCode.getText(node.value)}]`,
-            ),
-        }),
-      getExamples: (node) =>
-        ctx.report({
-          node,
+          node: cursor,
           messageId: "change",
           data: {
-            subject: "method",
-            from: "getExamples()",
-            to: ".meta()?.examples || []",
+            subject: "value",
+            from: "new DependsOnMethod(...)",
+            to: "its argument object and append its keys with ' /'",
           },
-          fix: (fixer) =>
-            fixer.replaceText(
-              node,
-              `(${ctx.sourceCode.getText(node.arguments[0])}.meta()?.examples || [])`,
-            ),
-        }),
+          fix: (fixer) => {
+            const makeMapper =
+              (feat?: "deprecated" | "nest") =>
+              (prop: TSESTree.ObjectLiteralElement) =>
+                isNamedProp(prop)
+                  ? `"${getPropName(prop)}${feat === "nest" ? "" : " /"}": ${ctx.sourceCode.getText(prop.value)}${feat === "deprecated" ? ".deprecated()" : ""},`
+                  : `${ctx.sourceCode.getText(prop)}, /** @todo migrate manually */`;
+            const nextProps = argument.properties
+              .map(makeMapper(isDeprecated ? "deprecated" : undefined))
+              .concat(nested?.properties.map(makeMapper("nest")) ?? [])
+              .join("\n");
+            return fixer.replaceText(cursor, `{\n${nextProps}\n}`);
+          },
+        });
+      },
     }),
 });
 
 export default {
-  rules: { v25 },
+  rules: { v26 },
 } satisfies TSESLint.Linter.Plugin;