+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 # Changelog
 
+## Version 26
+
+### v26.0.0
+
+- `DependsOnMethod` removed: use flat syntax with explicit method and a slash;
+
+```patch
+  const routing: Routing = {
+-   "/v1/users": new DependsOnMethod({
++   "/v1/users": {
+-     get: getUserEndpoint,
++     "get /": getUserEndpoint,
+-   }).nest({
+      create: makeUserEndpoint
+-   }),
++   },
+  };
+```
+
 ## Version 25
 
 ### v25.4.0